feat: Extract constants module (#102)
- Move __version__ and DEFAULT_VALUES to mgit/constants.py
- Add UpdateMode enum for standardized update modes
- Define configuration paths (CONFIG_DIR, CONFIG_FILE)
- Update __main__.py to import from constants module
