+++ mgit/constants.py
@@ -1,0 +1,25 @@
+"""Constants and default values for mgit."""
+from enum import Enum
+from pathlib import Path
+
+__version__ = "0.2.1"
+
+# Default values used if environment variables and config file don't provide values
+DEFAULT_VALUES = {
+    "AZURE_DEVOPS_ORG_URL": "https://www.visualstudio.com",
+    "LOG_FILENAME": "mgit.log",
+    "LOG_LEVEL": "DEBUG",
+    "CON_LEVEL": "INFO",
+    "DEFAULT_CONCURRENCY": "4",
+    "DEFAULT_UPDATE_MODE": "skip",
+}
+
+# Configuration paths
+CONFIG_DIR = Path.home() / ".config" / "mgit"
+CONFIG_FILE = CONFIG_DIR / "config"
+
+
+class UpdateMode(str, Enum):
+    skip = "skip"
+    pull = "pull"
+    force = "force"
